# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi API",
        "description": "Documentación de la API para mi aplicación",
        "contact": {
            "email": "friendhub@contact.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/exercises": {
            "get": {
                "summary": "Obtener todos los ejercicios",
                "description": "Devuelve una lista de todos los ejercicios disponibles",
                "operationId": "2127cfa4a1737fb80ac7a7abb55bbbb1",
                "responses": {
                    "200": {
                        "description": "Lista de ejercicios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "media": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "youtube_video_id": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear un nuevo ejercicio",
                "description": "Crea un nuevo ejercicio en la base de datos",
                "operationId": "60d93564a273a320c674e2993167efde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "youtube_video_id": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ejercicio creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            }
        },
        "/api/exercises/{id}": {
            "get": {
                "summary": "Obtener un ejercicio por ID",
                "description": "Devuelve un ejercicio específico por su ID",
                "operationId": "8e42fdf783928cdc6925db81fe4eff7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del ejercicio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "media": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "youtube_video_id": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ejercicio no encontrado"
                    }
                }
            },
            "put": {
                "summary": "Actualizar un ejercicio",
                "description": "Actualiza un ejercicio existente",
                "operationId": "65a19b1e7937962159dd0d451645df85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "youtube_video_id": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ejercicio actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Ejercicio no encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un ejercicio",
                "description": "Elimina un ejercicio específico por ID",
                "operationId": "20b6a40fe1d4eeef677718db223789cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ejercicio eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ejercicio no encontrado"
                    }
                }
            }
        },
...
    }
}